feature: 改造blocking与async桥接
